Split TLS implementation into its own translation unit.
SkTLS has it's own header separate from SkThread, and having SkThread own
the platform implementation of SkTLS is problematic with Chromium. The
simplest way to clean this up is to put the implementation in its own set
of files, where it is also more easily found.

R=robertphillips@google.com

Review URL: https://codereview.chromium.org/19240007

git-svn-id: http://skia.googlecode.com/svn/trunk/gyp@10105 2bbb7eff-a529-9590-31e7-b0007b416f81