Nobody defines SK_SUPPORT_HINTING_SCALE_FACTOR any more, so remove it.
I left the dummy value in the flattened paint.  I'm hoping to do an overhaul of
the paint serialization format which will need a picture version bump anyway,
so it can all go together.  If that doesn't pan out we can remove this on its
own later.

BUG=
R=reed@google.com

Review URL: https://codereview.chromium.org/23056005

git-svn-id: http://skia.googlecode.com/svn/trunk/gyp@10757 2bbb7eff-a529-9590-31e7-b0007b416f81